+++ database/migrations/2026_02_24_080248_add_batch_uuid_column_to_activity_log_table.php
@@ -8,15 +8,27 @@ class AddBatchUuidColumnToActivityLogTable extends Migration
 {
     public function up()
     {
-        Schema::connection(config('activitylog.database_connection'))->table(config('activitylog.table_name'), function (Blueprint $table) {
-            $table->uuid('batch_uuid')->nullable()->after('properties');
-        });
+        if (!Schema::connection(config('activitylog.database_connection'))->hasTable(config('activitylog.table_name'))) {
+            return;
+        }
+
+        if (!Schema::connection(config('activitylog.database_connection'))->hasColumn(config('activitylog.table_name'), 'batch_uuid')) {
+            Schema::connection(config('activitylog.database_connection'))->table(config('activitylog.table_name'), function (Blueprint $table) {
+                $table->uuid('batch_uuid')->nullable()->after('properties');
+            });
+        }
     }
 
     public function down()
     {
-        Schema::connection(config('activitylog.database_connection'))->table(config('activitylog.table_name'), function (Blueprint $table) {
-            $table->dropColumn('batch_uuid');
-        });
+        if (!Schema::connection(config('activitylog.database_connection'))->hasTable(config('activitylog.table_name'))) {
+            return;
+        }
+
+        if (Schema::connection(config('activitylog.database_connection'))->hasColumn(config('activitylog.table_name'), 'batch_uuid')) {
+            Schema::connection(config('activitylog.database_connection'))->table(config('activitylog.table_name'), function (Blueprint $table) {
+                $table->dropColumn('batch_uuid');
+            });
+        }
     }
 }
